+++ src/main/java/com/expensetracker/repository/SubcategoryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubcategoryRepository extends JpaRepository<Subcategory, Integer> {
+public interface SubcategoryRepository extends JpaRepository<Subcategory, Long> {
 
     Subcategory findByNameAndCategory(String name, Category category);
